// src/services/model/model/smsHomeRecommendProduct.ts
// 此文件由 oneapi-cli 自动生成，请不要编辑（This file is auto-generated by oneapi-cli, Please do not edit it）
/** */
export interface SmsHomeRecommendProduct {
  /** */
  id: number;
  /** */
  productId: number;
  /** */
  productName: string;
  /** */
  recommendStatus: number;
  /** */
  sort: number;
}